Python function transpose(m) that takes as input a two dimensional matrix m and returns the transpose of m.
A two dimensional matrix can be represented in Python row-wise, as a list of lists: each inner list represents one row of the matrix.
The transpose of a matrix converts each row into a column.
Write a Python function transpose(m) that takes as input a two dimensional matrix m and returns the transpose of m. The argument m should remain undisturbed by the function.